requirements.txt
langchain
langchain-core
langchain-community
langchain-huggingface
langgraph
python-dotenv
streamlit
langgraph-checkpoint-sqlite
langsmith
langchain-google-genai
google-generativeai

